modules created
PSFModel.py contains the main()
