+++ yapsfm_test/zernike.py
@@ -1,0 +1,76 @@
+#! /usr/bin/env python
+"""
+Contains functions to compute Zernike modes.
+
+Inputs:
+    n: mode's radial
+    m: mode's asimuthal
+    r: a meshgrid object of radial positions
+    theta: a meshgrid object of asimuthal positions
+
+Outputs:
+    an array giving the shape of the wavefront
+"""
+
+import numpy as np
+
+
+def even_zernike(n, m, r, theta):
+    """
+    Computes the even Zernike mode following Noll (1976)
+
+    :param n: radial order of the mode
+    :type n: int
+    :param m: azimuthal order of the mode
+    :type m: int
+    :param r: radial position
+    :type r: float
+    :param theta: angle
+    :type theta: float
+    :return: the value of the Zernike mode corresponding to n and m, at position (r,theta).
+    :rtype: float
+    """
+
+    zernike_value = np.sqrt(n+1)*r_nm(n, m, r)*np.sqrt(2)*np.cos(m*theta)
+    return zernike_value
+
+
+def odd_zernike(n, m, r, theta):
+    """
+    Computes the odd Zernike mode following Noll (1976)
+
+    :param n: radial order of the mode
+    :type n: int
+    :param m: azimuthal order of the mode
+    :type m: int
+    :param r: radial position
+    :type r: float
+    :param theta: angle
+    :type theta: float
+    :return: the value of the Zernike mode corresponding to n and m, at position (r,theta).
+    :rtype: float
+    """
+
+    zernike_value = np.sqrt(n+1)*r_nm(n, m, r)*np.sqrt(2)*np.sin(m*theta)
+    return zernike_value
+
+
+def r_nm(n, m, r):
+    """
+    Computes the radial part R_n^m(r), with r a meshgrid object
+
+    :param n: radial order of the polynomial
+    :type n: int
+    :param m: azimuthal order of the polynomial
+    :type m: int
+    :param r: radial position
+    :type r: float
+    :return: radial part of the Zernike mode at position r
+    :rtype: float
+    """
+
+    r_nm_value = 0
+    for s in range((n-m)/2+1):
+        r_nm_value += (((-1)**s * np.math.factorial(n-s)) / (np.math.factorial(s) * np.math.factorial((n+m)/2-s) *
+                                                             np.math.factorial((n-m)/2-s))) * r**(n-2*s)
+    return r_nm_value
